The file **milestone_2.py** is a python file that holds a list of my favourite fruit in the variable word_list. a random word is taken fromm the word_list using the method random.choice and stored in the variable word. The user is asked to put in a single character which is stored in the variable guess.
An if statement is used to check the length is 1  and whether the charcter entered is an alpha.If it is then good guess! statement is printed out.

If this is not the case the following else statment is triggered and Oops! That is not a valid input. is printed.