## milestone_3.py
The file **milestone_3.py** is a upgraded version of **mileston_2.py**. Here 2 functions have been produced.
*check_guess* takes in the gussed letter as an argument and checks if the letter is in the word that the computer had randomly chose from the word_list containing a list of 5 fruits.
the function *ask_for_input* is the main driver of the programme. IT has a while loop set to true. It then asks for input from the user. then there is an if statement that checks that the input is valis i.e.
the length is 1 and it is an alphabet character. If it passes this test the while loop is broken. if it fails then a print statement asks for the user to try again.