##milestone_4.py
this file is a possible example of oop programming. A class Hangman is created and contains everything for playing a round of hangman. In this class the following variables are intialised.
the following variables are definded
the word_list and number _of Lives(which is set to 5) are passed as arguemnts and are set as self. i.e. self.word_list and self_num_of_lives.
the word is chosen using the random.choice function and called self.word.
the list word_guessed is initalised using '' to represent each letter. The list is as long as the word.
the num_letters is the number of UNIQUE letters in the word using (list(len(set.self.word)).
an empty list called list_of _guesses.

the check_guess function is expanded from mileston_3.py.
 a for loop is added in the if statement after the print statement. the for loop looks at each letter(variable each_letter) and the posistion(variable count) it holds in the word. it then compares it to the guess and if it is the same then places the letter in the list word_guessed at the index count. Outside the if statement containing the for loop  but within the  outside if statement the num_letters is reduced by 1.
 The else statement is also changed.
 num_lives is reduced by 1
 an additional print statement is added to let the user know that there are a number of lives left.