changes to milestone_4.py
A new function is created called play_game . It takes word_list as an argument. Ouside of this function an instance of Hangman is created called game. Inside play_game a while loop is set to True. inside the while loop an if statement checks if the num_of_lives is 0. If it is then the player is informed that they have lost. The loop is broken here. If the num_of_lives is greater than 0 and the num_letters is 0 then the user is informed that they have won. The loop si broken here. IF the num_of_lives is greater than 0 then the ask_input method in game is triggered and the user is asked to input a letter etc.
 The only change to the methods described in milestone_4.py is that thewhile loop in ask_for_input is no longer required as it is now in play_game.